benchmarks: gpio-echo: external latency tester tool
This Linux-side tool is designed to echo the GPIO signals issued from
a remote board, which in turn measures the response time. The code has
been validated using a Linux-based rpi3b running gpio-echo, and a
Zephyr-based FRDMk64f board measuring the response time.

The current version of the Zephyr code is available in the zephyr/
subdirectory in patch format. This code is maintained at:
https://github.com/ldts/zephyr.git, branch evl-latency

For instance, GPIO23 can be used to receive test signals on rpi3b and
GPIO24 to respond to them.

gpio-echo is realtime capable (via EVL) or non-realtime capable
(standard Linux behaviour) - it depends on how you run it.

$ gpio-echo -n gpiochip0 -o 23 -t 24 -O -T -f

Once that process is started (and the necessary cabling is done) start
this Zephyr program, get the console and follow the instructions.

Connections:
-------------
 Zephyr - FRDMk64F:                Linux - rpi3b
  PIN 20 (PTE-24) ----------------  PIN 16 (GPIO 23)
  PIN 18 (PTE-25) ----------------  PIN 18 (GPIO 24)

Signed-off-by: Jorge Ramirez-Ortiz <jro@xenomai.org>