lib: add support for event flag groups
Event flags are boolean conditions represented by a single bit,
grouped in 32bit words. All bits from a group are initially set to
zero.

An event flag group is a lightweight notification mechanism. The
application can send bitmasks to raise individual bits from the group
(i.e. group_value |= bits), or wait for the group to have at least one
bit set for satisfying the request. In the latter case, the group
value is read then cleared atomically, and the collected bits are
returned to the thread heading the wait queue.