lib: normalize return values of evl_new_*(), evl_open_*() calls
Some creation calls have to return the file descriptor of the new
element because there is no in-memory descriptor associated with the
latter where we could stick the former. So for consistency, just have
all evl_new_*() calls return the file descriptor referring to the new
element.

Likewise, all evl_open_*() calls now do the same.