lib/proxy: enable fixed granularity for writes
This change allows the application to specify a fixed granularity for
writes to the target file, such as connecting the proxy with an
eventfd for instance, guaranteeing fixed-size 64bit writes to the
latter.

evl_poll() is also enabled for the proxy, with POLLOUT|POLLWRNORM
raised when the output has drained, i.e. all data have been sent to
the target file.