lib/init: better detect ABI discrepancies
Post ABI 18 libevl would fail to detect incompatible core support by
receiving ENOTTY as a result of asking for the core information, due
to the change in size of the argument struct which was introduced to
support ABI ranges. Make sure to trap ENOTTY on this call as an ABI
mismatch too.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>