include/evl: fix unresolved references to evl_sched_attrs.sched_priority
sched_priority is implemented as a macro by glibc which serves as a
wrapper to the real attribute field in schedparam. Some architectures
may use this trick to hide the actual identifier from the user code,
so make sure to always pull <sched.h> before referencing
sched_priority in uapi/ headers, so that such wrapping also happens
when struct evl_sched_attrs is defined.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>