benchmarks/latmus: add sirq latency measurement
This mode measures the delay between the moment a synthetic interrupt
is posted from the oob stage and when it is eventually received by its
in-band handler. When measured under significant pressure, this gives
the typical interrupt latency experienced by the in-band kernel due to
local interrupt disabling.

Therefore, this is an in-band only test which measures IRQ latency
experienced by the common kernel infrastructure, _NOT_ by EVL.

Measurement is requested with '-s' option.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>