utils: checklist: mark CONFIG_DEBUG_ENTRY as latency hog
Kernel entry instrumentation/debug code has to run in the hot path by
definition, add it to the list of settings which may increase
latency. This feature is only available for x86 as of kernel 5.8, but
it is likely going to spread all over the place in the future, so
register it as generic.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>