lib/init: factor out generic vDSO lookup task
There is no point in specializing the resolution of clock_gettime() in
the vDSO on a per-arch basis. This is mostly generic code with only a
couple of arch-specific identifiers we need, which we can leave into
the arch/ area.

Move the bulk of the symbol resolution code to the generic init
routine, adding asm/evl/vdso.h for storing the few arch-specific bits
such code still depends on.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>